refactor: 리뷰 평점 필드 int 타입으로 수정 # 232

+++ backend/fittoring/src/main/java/fittoring/mentoring/business/service/dto/ReviewModifyDto.java
@@ -5,7 +5,7 @@ import fittoring.mentoring.presentation.dto.ReviewModifyRequest;
 public record ReviewModifyDto(
     Long reviewerId,
     Long reviewId,
-    byte rating,
+    int rating,
     String content
 ) {
