+++ backend/fittoring/src/main/java/fittoring/mentoring/presentation/dto/MentoringReviewGetResponse.java
@@ -3,7 +3,7 @@ package fittoring.mentoring.presentation.dto;
 public record MentoringReviewGetResponse(
     Long id,
     String reviewerName,
-    byte rating,
+    int rating,
     String content
 ) {
